requirements.txt
mysqlclient==1.3.10
Django==1.10.5
django-bootstrap-themes==3.3.6
django-bootstrap3==8.2.2django-registration-redux==1.4